+++ lib/continuum/pure.ex
@@ -25,7 +25,7 @@ defmodule Continuum.Pure do
 
   @doc false
   def __on_definition__(env, _kind, _name, _args, _guards, body) when not is_nil(body) do
-    case Continuum.AstCheck.scan(body, env.file) do
+    case Continuum.AstCheck.scan(body, env) do
       :ok ->
         :ok